+++ pynenc/runner/__init__.py
@@ -1,5 +1,6 @@
 from pynenc.runner.base_runner import BaseRunner, DummyRunner
 from pynenc.runner.multi_thread_runner import MultiThreadRunner
+from pynenc.runner.persistent_process_runner import PersistentProcessRunner
 from pynenc.runner.process_runner import ProcessRunner
 from pynenc.runner.thread_runner import ThreadRunner
 
@@ -9,4 +10,5 @@ __all__ = [
     "ThreadRunner",
     "ProcessRunner",
     "MultiThreadRunner",
+    "PersistentProcessRunner",
 ]